backend/requirements.txt
Django==6.0.3
djangorestframework==3.16.0
django-cors-headers==4.7.0
psycopg[binary]==3.2.9
